requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
anyio==3.6.1
argcomplete==2.0.0
async-timeout==4.0.2
asyncclick==8.1.3.2
attrs==21.4.0
awscli==1.23.13
botocore==1.25.13
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
distlib==0.3.4
docutils==0.15.2
dominate==2.6.0
filelock==3.7.0
Flask==2.1.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-JWT-Extended==4.4.1
Flask-Login==0.6.0
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
marshmallow==3.16.0
multidict==6.0.2
numpy==1.22.4
packaging==21.3
pipenv==2022.5.2
pipx==1.1.0
platformdirs==2.5.2
pyasn1==0.4.8
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-kasa==0.5.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.0
rsa==4.7.2
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
tplink-cloud-api==4.1.3
twilio==7.9.3
typing_extensions==4.2.0
urllib3==1.26.9
userpath==1.8.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
visitor==0.1.3
Werkzeug==2.1.1
WTForms==3.0.1
yarl==1.7.2
zipp==3.8.0
